Исправления для карты и формы
